Fix compilation when wint_t is unsigned.
We should figure out a better mechanism to address this issue. This feels more
like a bandaid than a real fix.

+++ src/command_with_modifiers.cc
@@ -23,7 +23,7 @@ class CommandWithModifiersMode : public EditorMode {
 
   void ProcessInput(wint_t c, EditorState* editor_state) {
     buffer_->Undo(editor_state, OpenBuffer::ONLY_UNDO_THE_LAST);
-    switch (c) {
+    switch (static_cast<int>(c)) {
       case '\n':
       case ' ':
         RunHandler(editor_state, CommandApplyMode::FINAL);
